fs-util: Make xopenat() take flags as int64_t
This gives us some space to define our own flags.

+++ src/basic/fs-util.h
@@ -132,6 +132,6 @@ int open_mkdir_at(int dirfd, const char *path, int flags, mode_t mode);
 
 int openat_report_new(int dirfd, const char *pathname, int flags, mode_t mode, bool *ret_newly_created);
 
-int xopenat(int dir_fd, const char *path, int flags, mode_t mode);
+int xopenat(int dir_fd, const char *path, int64_t flags, mode_t mode);
 
-int xopenat_lock(int dir_fd, const char *path, int flags, mode_t mode, LockType locktype, int operation);
+int xopenat_lock(int dir_fd, const char *path, int64_t flags, mode_t mode, LockType locktype, int operation);
